@@ -2,30 +2,29 @@
   "name": "h5webapp-vite",
   "version": "1.0.0",
   "private": true,
-  "description": "Example of using @h5web/app with Vite.",
-  "keywords": [
-    "hdf5",
-    "visualization",
-    "heatmap",
-    "h5web"
-  ],
+  "packageManager": "pnpm@9.4.0",
+  "type": "module",
   "scripts": {
     "dev": "vite",
+    "prebuild": "pnpm \"/^lint:/\"",
     "build": "tsc && vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "lint:tsc": "tsc",
+    "lint:prettier": "prettier --cache --check ."
   },
   "dependencies": {
-    "@h5web/app": "6.1.0",
+    "@h5web/app": "12.0.0",
     "normalize.css": "8.0.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "react": "18.3.1",
+    "react-dom": "18.3.1"
   },
   "devDependencies": {
-    "@types/node": "^16.0.0",
-    "@types/react": "17.0.33",
-    "@types/react-dom": "17.0.10",
-    "@vitejs/plugin-react": "1.3.2",
-    "typescript": "4.7.4",
-    "vite": "2.9.14"
+    "@types/node": "^20.14.9",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@vitejs/plugin-react": "4.2.1",
+    "prettier": "^3.3.2",
+    "typescript": "5.5.2",
+    "vite": "5.3.1"
   }
 }
