@@ -1,7 +1,11 @@
-import ReactDOM from 'react-dom';
-
 import './index.css';
+
+import { assertNonNull } from '@h5web/app';
+import { createRoot } from 'react-dom/client';
 
 import MyApp from './MyApp';
 
-ReactDOM.render(<MyApp />, document.getElementById('root'));
+const rootElem = document.querySelector('#root');
+assertNonNull(rootElem);
+
+createRoot(rootElem).render(<MyApp />);
